i965/fs: Add support for gl_Layer input
This is a bit unusual -- we want to treat it as an input on the GLSL
side rather than a `system value` oddball. However, it doesn't get a
real slot in the VUE layout -- earlier shader stages write it into the
vertex header, and it appears in the FS payload packed into some R0.0
bits.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>