i965/Gen7: Fix HiZ ops for MSAA depth
Previously, we would program the sample count based on the blorp
operation's dst sample count -- which is zero for a HiZ op; we'd also
ignore the difference between physical and logical surface size for
the depth surface.

That kindof worked -- the two errors almost cancel out -- but produced
strange blocky artifacts.

Instead, program the sample count properly, and use the logical
dimensions of the depth surface.

Fixes broken rendering in `Tesseract` with msaa=4 glineardepth=0

NOTE: I've hacked the BLORP depth alignment code a bit (at `Not quite
sure ...`). This is almost certainly not the right thing.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>