Add prefer_no_swizzle parameter to ChooseTextureFormat
This is a hint that we want components ordered as they are in the
internalformat name.

We want this for views (and so also for
immutable-format textures from which views might be created).

In other cases we might want to prefer ARGB or BGRA component order
to avoid reswizzling uploads, or for interaction with scanout hardware.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>