swrast: Defer SpanArrays allocation until actually necessary.
Avoids allocating about 16M of stuff in cases where we never actually do
any rendering with swrast, but need to create a context for possible fallbacks.

Slightly suspect thing is that this moves a large allocation into a path
which is not really expecting failures. We emit GL_OUT_OF_MEMORY in this
case, but then will very likely segfault and die in a heap.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>