i965: Suppress re-flagging of BRW_NEW_UNIFORM_BUFFER while emitting UBO state.
This shouldn't actually happen in a sensible app (the easiest way to
force it is to use a UBO in a shader without binding anything to the
binding point). When it *does* happen though, it makes a loop in the
dirty state, which causes an assertion failure in the state validator if
any debug flag is enabled.

[This is almost certainly not the right way to fix it, but maybe the
patch will get a discussion going about excessive flagging...]

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>