i965/hsw: Properly handle RO state invalidation
Before invalidating the "RO state" caches, we have to emit a PIPE_CONTROL...

The following is the best I can explain things as I understand them without
getting too far into the low-level details.

I believe the "read only" aspect of this is misleading. It just so happens that
the state caches are read only, but really this is about state cache coherency.
As I understand it, when invalidating any of the state caches, the internal
hardware that handles this stuff (state arbiter, sometimes referred to as sarb)
must flush and sync on any pending requests before it can send an invalidation.
Why this is needed is based on the uArch design. Essentially you can have
multiple pending requests for *things* and the invalidation must occur after
those pending things complete.

The result of this is the following:
   Programming Note: PIPECONTROL with RO Cache Invalidation: Prior to
   programming a PIPECONTROL command with any of the RO cache invalidation bit
   set, program a PIPECONTROL flush command with “CS stall” bit and “HDC Flush”
   bit set.

Unfortunately, there is no such thing as an "HDC Flush" bit anywhere that I can
find. AFAICT, HDC is short for H? Data Cluster. So I've settled on
PIPE_CONTROL_DATA_CACHE_INVALIDATE as the closest sounding thing.

I've put this code before the existing every 4 PIPE_CONTROL workaround because
this might emit a PIPE_CONTROL which would void the need to use that workaround.
It probably makes no difference since back to back CS stalls probably have
little measurable impact anyway - but it would be less commands.

v2: Rebased on master
Updated commit message
Filled picked a different PIPE_CONTROL flag

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>