i965: Reorganize PMA optimization workaround code
Renaming functions are useful for having a more generic set of calls to make for
different types of PMA optimizations (Gen9 brings a new one). Centralizing the
functions will help with upcoming SKL work in the same area. For instance, the
current workaround we have is supposedly not needed on SKL (next patch). There
are no intended functional changes here.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>